qt6: move memento to qt6
This is the first attempt to move Memento over to Qt6. As much as I'd
like this to happen ASAP, and I have no reason to believe that Memento
will be unstable with these changes, the problem is that KDE Plasma does
not yet support Qt6. This means that Breeze does not yet support Qt6,
which means Qt falls back to bundled themes (which look like shit). This
doesn't affect macOS or Windows users, but it does affect me personally.
Therefore I refuse to push this to master until Qt6 support is released
for Breeze in the year 2049.

Of note is the changes made to SubtitleListWidget. Previously I depended
on QList allocating memory in the heap for me and keeping track of it
all on its own. Now that QList is just QVector and no longer does heap
allocation of elements greater than 8 bytes, I can no longer depend on
this behavior. Therefore I have just started copying SubtitleInfos
because that's the easiest way to do things and they aren't that large
anyway.